Create reports sub-package within internal and refactor
