+++ internal/helpers.go
@@ -28,8 +28,8 @@ type Counter struct {
 
 var Cntr Counter
 
-// contains check string present in list.
-func contains(l []string, str string) bool {
+// Contains check string present in list.
+func Contains(l []string, str string) bool {
 	for _, s := range l {
 		if s == str {
 			return true
